services/preprocessor/pyproject.toml
[project]
name = "preprocessor"
version = "0.1.0"
description = "Preprocessor service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.120.0",
    "uvicorn[standard]>=0.35.0",
    "kafka-python>=2.2.0",
    "prometheus-client",
    "python-multipart",
    "numpy",
    "pillow",
    "huggingface-hub"
]